# util/mon: add optional threadsafety to BoundAccount