@dt 2022-01-12
> Should we add a MakeBoundAccountWithMu?

Sure we could. I'd just exported it so we could manually set it before entering concurrent call trees and then potentially unset it after if we wanted, e.g. if we were borrowing someone else's account.