@adityamaru 2022-01-12
> if we were borrowing someone else's account.

Makes sense, lgtm as is then!