@jordanlewis 2022-01-14
I'm not convinced that this is the right change. The BoundAccount abstraction is very clear today and I think that adding a mutex, and making the abstraction multi-purpose, muddies the waters and will change how the API is used and introduce extra complexity into what's been a very long-lasting and solid interface.

I don't fully understand your argument about needing an interface to have a locking implementation. We could create a `LockableBoundAccount` that wraps a `BoundAccount`, and pass that around in code that needs it. Most code should not need the extra lock.

Alternatively, I would prefer to see us create abstractions around that "re-merging" pattern you mentioned in the thread. The pattern you showed seems very sound and reusable. Why not encapsulate that pattern in a new object and package it up for reuse?

And finally, I'm actually confused overall about why you need to do this. BoundAccounts ladder up to monitors already. Can you use a monitor to keep track of the overall memory usage of the many threads, rather than creating a tree of BoundAccounts?