@dt 2022-01-15
>  I'm assuming that there is a coordinator goroutine

Yep, there is -- we can indeed just not close the sub-accounts in each goroutine, but the part that seems to complicate that is that the lifetimes of the reservations in those accounts aren't the same as the root account -- we need to be able to go back and shrink some of those reservations back out and we'd need to shrink from the right one, right? 

If we went that route, I think I'd probably opt o instead add a `MergeAndCloseSiblingAccount` to BoundAccount, which can be passed another BA and takes over reservations (after asserting it has the same monitor parent)? Then we'd be able to just merge them all in to the root account and then just shrink it as needed. I guess this is helper-to-make-many-monitors approach.

I dunno, I think I still like just making it thread safe but can let this marinate for a while -- there's no real urgency here.