@yuzefovich 2022-01-18
> we need to be able to go back and shrink some of those reservations back out and we'd need to shrink from the right one, right?

Yep, makes sense.

> I dunno, I think I still like just making it (optionally) thread safe but can let this marinate for a while -- there's no real urgency here.

I agree, the current approach seems cleaner, LGTM.